Compiler, startup and daemon termination changes.
The compiler in the makefile has changed over to be gcc rather than
clang. No real preference here. The command line interface for the
program will now allow no parameters to be supplied which will default
its execution in RC_STANDALONE mode. The server process is now
explicitly terminating at the end of the main server routine as I was
noticing zombies gathering in the process table after subsequent runs.